@bencrts 2024-12-18
@malb are you against using `numpy` in general? I think I can also speed-up the simulator/s using a similar trick.