@malb 2024-12-18
NumPy only does double precision, right? If that's all we need then I think it makes sense to call it (given that it is shipped with Sage anyways)